disgusting but it's progress
yes, this is full of loads of llm comments and the same functions are
copy/pasted and commented out 3 times in a row. but it's late and i just
spent hours tracking down a bug with how packets were being processed
and it's finally working exactly as intended! so i'm just slamming the
commit button and heading to bed