send motd
tbh not actually clear if this is working. there are several screens
that show blank text on the emulator and only render correctly if you
use the software renderer. on the motd screen however, it's blank
regardless of which renderer is used. unclear to me if this is due to a
malformed motd packet or if it's the client's fault.