many bug fixes & start of GSpackethandler
where the java code exclusively relies on getters & setters, i've been
conservative in my use. unfortunately that led to situations where the
raw data needed to be transformed before it was returned but i was
accessing the raw data directly.

tbh the code should probably be refactored to use getters and setters
across the board but that's a task for another day!